+++ src/bsp/extrem.ml
@@ -2,8 +2,7 @@ open Graphics
 open Geometry
 open Bsp
 
-module Bsp_extrem : Bsp_type = struct
-  let min_area = 10000.
+module Make (C : Settings.Colors) : Bsp_type = struct
 
   type bsp = R of region_label | L of line_label * bsp * bsp
 
@@ -40,11 +39,15 @@ module Bsp_extrem : Bsp_type = struct
          else left, change_color ~reverse:reverse right pt
        in
        L (l, left, right)
-    | R region -> R {region_id=region.region_id;region_color=(next_color reverse region.region_color)}
+    | R region ->
+       R {
+           region_id = region.region_id;
+           region_color= (C.next_color reverse region.region_color)
+         }
 
   let rec nb = ref 0
 
-  and add_random_line localBsp pts localDepth =
+  and add_random_line localBsp pts localDepth min_area =
     if localDepth = 0
     then localBsp
     else
@@ -52,9 +55,9 @@ module Bsp_extrem : Bsp_type = struct
       | L (l, left, right) ->
          let leftPts, rightPts = split_by_line l.section pts
          in if Random.float 1. < 0.5
-            then L(l, add_random_line left leftPts (localDepth - 1), right)
-            else L(l, left, add_random_line right rightPts (localDepth - 1))
-      | R _ ->
+            then L(l, add_random_line left leftPts (localDepth - 1) min_area, right)
+            else L(l, left, add_random_line right rightPts (localDepth - 1) min_area)
+      | R c ->
          let pts = List.sort (compare_counter_clockwise (center pts)) pts in
          let ptsArr = Array.of_list pts in
          let new_line = gen_random_lines ptsArr in
@@ -68,25 +71,25 @@ module Bsp_extrem : Bsp_type = struct
          then begin
              nb := !nb + 1;
              L({line_color = 0; line_id = 0; section = new_line},
-                R {region_color = rand_color (); region_id = 0},
-                R {region_color = rand_color (); region_id = 0})
+                R {region_color = C.rand_color (); region_id = 0},
+                R {region_color = C.rand_color (); region_id = 0})
            end
          else localBsp
 
-  and gen_random_bsp width height nb_lines maxDepth =
+  and gen_random_bsp width height nb_lines maxDepth min_area =
     if maxDepth >= 0 && (2. ** (float_of_int maxDepth)) >= (float_of_int nb_lines /. 2.)
-    then R {region_id = 0; region_color = rand_color ()}
+    then R {region_id = 0; region_color = C.rand_color ()}
     else
       let bsp = ref (R {region_id = 0; region_color = white}) in
       let v, _ = edges width height in
       for _ = 1 to nb_lines do
-        bsp := add_random_line !bsp v maxDepth
+        bsp := add_random_line !bsp v maxDepth min_area
       done;
       !bsp
 
-  and generate_random_bsp width height =
-    let bsp = gen_random_bsp width height 100 (-1) in
+  and generate_random_bsp width height min_area =
+    let bsp = gen_random_bsp width height 100 (-1) (float_of_int min_area) in
     bsp, !nb
 end
 
-module Bsp = Bsp.Make (Bsp_extrem)
+module Bsp (S : Settings.Game_settings) (C : Settings.Colors) = Bsp.Make(S)(Make(C))